# optimize Neg