```
goos: darwin
goarch: arm64
pkg: github.com/shogo82148/int128
               │   .old.txt   │              .new.txt               │
               │    sec/op    │    sec/op     vs base               │
Int128_Neg-10    0.6225n ± 1%   0.6254n ± 1%       ~ (p=0.055 n=10)
Uint128_Neg-10   0.6331n ± 1%   0.6251n ± 0%  -1.26% (p=0.000 n=10)
geomean          0.6278n        0.6253n       -0.40%
```